requirements/requirements.txt
svgwrite==1.4
argparse==1.4.0
pre-commit==2.7.1
black==20.8b1
cairosvg==2.5.0
matplotlib==3.3.2
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
natsort==3.3.0
